Access tiles from ArcGIS Online and ArcGIS Server to visualize and identify features.
Source: R/tiledMapLayers.R

If you have published a Feature Service in ArcGIS Online, it can be used to create a static set of tiles as well. You can find details about that process in the ArcGIS Online Help. Your map service must be published using the Web Mercator Auxiliary Sphere tiling scheme (WKID 102100/3857) and the default scale options used by Google Maps, Bing Maps and ArcGIS Online. Esri Leaflet will not support any other spatial reference for tile layers.

addEsriTiledMapLayer(map, url, options = tiledMapLayerOptions(),
  layerId = NULL, group = NULL)

Arguments

map
The leaflet map

url
URL of the Map Service with a tile cache.

options
options for the tiledmap layer.

layerId
A unique ID for the layer.

group
The name of the group this layer should be added to.

Examples

leaflet() %>% setView(-81.47, 30.70, 12) %>%
  addEsriTiledMapLayer(
    url = "https://services.arcgisonline.com/ArcGIS/rest/services/USA_Topo_Maps/MapServer")

## for more examples see
# browseURL(system.file("examples/tiledMapLayers.R", package = "leaflet.esri"))