Add Geodesic Lines

Add Geodesic Lines

addGeodesicPolylines(map, lng = NULL, lat = NULL, layerId = NULL,
  group = NULL, steps = 10, wrap = TRUE, stroke = TRUE,
  color = "#03F", weight = 5, opacity = 0.5, dashArray = NULL,
  smoothFactor = 1, noClip = FALSE, popup = NULL, popupOptions = NULL,
  label = NULL, labelOptions = NULL, options = pathOptions(),
  highlightOptions = NULL, data = getMapData(map))

addGreatCircles(map, lat_center = NULL, lng_center = NULL, radius,
  layerId = NULL, group = NULL, steps = 10, wrap = TRUE,
  stroke = TRUE, color = "#03F", weight = 5, opacity = 0.5,
  dashArray = NULL, smoothFactor = 1, noClip = FALSE, popup = NULL,
  popupOptions = NULL, label = NULL, labelOptions = NULL,
  options = pathOptions(), highlightOptions = NULL,
  data = getMapData(map))

Arguments

map

map object

lng

vector of longitudes

lat

vector of latitudes

layerId

the layer id

group

the name of the group this raster image should belong to (see

steps

Defines how many intermediate points are generated along the path. More steps mean a smoother path.

wrap

Wrap line at map border (date line). Set to 'false' if you want lines to cross the dateline (experimental, see noWrap-example on how to use)

stroke

whether to draw stroke along the path (e.g. the borders of polygons or circles)

color

stroke color

weight

stroke width in pixels

opacity

stroke opacity (or layer opacity for tile layers)

dashArray

a string that defines the stroke https://developer.mozilla.org/en/SVG/Attribute/stroke-dasharray

smoothFactor

how much to simplify the polyline on each zoom level

noClip

whether to disable polyline clipping (more means better performance and less accurate representation)

popup

a character vector of the HTML content for the popups (you are recommended to escape the text using htmlEscape()

popupOptions

A Vector of popupOptions to provide popups for security reasons)

label

a character vector of the HTML content for the labels

labelOptions

A Vector of labelOptions to provide label options for each label. Default NULL

options

a list of additional options, intended to be provided by a call to pathOptions()

highlightOptions

Options for highlighting the shape on mouse over.

data

map data

lat_center, lng_center

lat/lng for the center

radius

in meters

Functions

addGreatCircles: Adds a Great Circle to the map