removes the geojson choropleth.

removes the geojson choropleth.

clears the geojson choropleth.

addGeoJSONChoropleth(map, geojson, layerId = NULL, group = NULL,
  valueProperty, labelProperty, popupProperty, scale = c("white", "red"),
  steps = 5, mode = "q", colors = NULL, stroke = TRUE, color = "#03F",
  weight = 1, opacity = 0.5, fillOpacity = 0.2, dashArray = NULL,
  smoothFactor = 1, noClip = FALSE)

removeGeoJSONChoropleth(map, layerId)

clearGeoJSONChoropleth(map)

Arguments

map

The leaflet map

geojson

either as a list or a string

layerId

a unique ID for the layer

group

the group this layer will be added to

valueProperty

The property to use for coloring

labelProperty

The property to use for labelling.

popupProperty

The property to use for popup content

scale

The scale to use from chroma.js

steps

number of breakes

mode

q for quantile, e for equidistant, k for k-means

colors

overrides scale with manual colors

stroke

whether to draw stroke along the path (e.g. the borders of polygons or circles)

color

stroke color

weight

stroke width in pixels

opacity

stroke opacity (or layer opacity for tile layers) circles)

fillOpacity

fill opacity

dashArray

a string that defines the stroke https://developer.mozilla.org/en/SVG/Attribute/stroke-dasharray

smoothFactor

how much to simplify the polyline on each zoom level (more means better performance and less accurate representation)

noClip

whether to disable polyline clipping