Options for the GroupedLayersControl

Options for the GroupedLayersControl

Usage

groupedLayersControlOptions(
  exclusiveGroups = NULL,
  groupCheckboxes = TRUE,
  groupsCollapsable = TRUE,
  groupsExpandedClass = "leaflet-control-layers-group-collapse-default",
  groupsCollapsedClass = "leaflet-control-layers-group-expand-default",
  sortLayers = FALSE,
  sortGroups = FALSE,
  sortBaseLayers = FALSE,
  collapsed = TRUE,
  autoZIndex = TRUE,
  ...
)

Arguments

exclusiveGroups

character vector of layer groups to make exclusive (use radio buttons)

groupCheckboxes

Show a checkbox next to non-exclusive group labels for toggling all

groupsCollapsable

Should groups be collapsible? Default is TRUE

groupsExpandedClass

The CSS class of expanded groups

groupsCollapsedClass

The CSS class of collapsed groups

sortLayers

Sort the overlay layers alphabetically? Default is FALSE

sortGroups

Sort the groups alphabetically? Default is FALSE

sortBaseLayers

Sort the baselayers alphabetically? Default is FALSE

collapsed

if TRUE (the default), the layers control will be rendered as an icon that expands when hovered over. Set to FALSE to have the layers control always appear in its expanded state.

autoZIndex

if TRUE, the control will automatically maintain the z-order of its various groups as overlays are switched on and off.

...

other options for layersControlOptions()

See also