leaflet.extras: Extra Functionality for 'leaflet' Package.

The 'leaflet' JavaScript library provides many plugins some of which are available in the core 'leaflet' package, but there are many more. It is not possible to support them all in the core 'leaflet' package. This package serves as an add-on to the 'leaflet' package by providing extra functionality via 'leaflet' plugins.

See also

Author

Maintainer: Sebastian Gatscha sebastian_gatscha@gmx.at

Authors:

Bhaskar Karambelkar bhaskarvk@gmail.com

Barret Schloerke barret@rstudio.com

Other contributors:

Bangyou Zheng Bangyou.Zheng@csiro.au (Leaflet-search and Leaflet-GPS plugin integration) [contributor]

Robin Cura robin@cura.info (Fixes for Draw Options) [contributor]

Markus Voge markus.voge@ea-aw.de (Enhancements for Draw Options) [contributor]

Markus Dumke (Bounce Marker addition) [contributor]

Mapbox (leaflet-omnivore, csv2geojson, and togeojson libraries) [contributor, copyright holder]

Henry Thasler (Leaflet.Geodesic library) [contributor, copyright holder]

Dennis Wilhelm (Leaflet.StyleEditor library) [contributor, copyright holder]

Kirollos Risk (fuse.js library) [contributor, copyright holder]

Tim Wisniewski (leaflet-choropleth library) [contributor, copyright holder]

Leaflet (leaflet-draw library) [contributor, copyright holder]

Alexander Milevski (leaflet-draw-drag library) [contributor, copyright holder]

John Firebaugh (leaflet-fullscreen library) [contributor, copyright holder]

Stefano Cudini (leaflet-gps library) [contributor, copyright holder]

Johannes Rudolph (leaflet-hash library) [contributor, copyright holder]

Per Liedman (leaflet-measure-path library) [contributor, copyright holder]

Pavel Shramov (leaflet-plugins library) [contributor, copyright holder]

Filip Zavadil (leaflet-pulse-icon library) [contributor, copyright holder]

Stefano Cudini (leaflet-search library) [contributor, copyright holder]

CliffCloud (leaflet-sleep library) [contributor, copyright holder]

Ursudio (leaflet-webgl-heatmap library) [contributor, copyright holder]

Maxime Hadjinlian (leaflet.BounceMarker library) [contributor, copyright holder]

Vladimir Agafonkin (leaflet.heat library) [contributor, copyright holder]

Iván Sánchez Ortega (leaflet.tilelayer.pouchdbcached library) [contributor, copyright holder]

Dale Harvey (pouchdb-browser library) [contributor, copyright holder]

Mike Bostock (topojson library) [contributor, copyright holder]