Make pulse-icon set

Make pulse-icon set

Make Pulse Icon

An icon can be represented as a list of the form list(color, iconSize, ...). This function is vectorized over its arguments to create a list of icon data. Shorter argument values will be re-cycled. NULL values for these arguments will be ignored.

Add Pulse Markers

pulseIconList(...)

# S3 method for leaflet_pulse_icon_set
[(x, i)

makePulseIcon(color = "#ff0000", iconSize = 12, animate = TRUE,
  heartbeat = 1)

pulseIcons(color = "#ff0000", iconSize = 12, animate = TRUE,
  heartbeat = 1)

addPulseMarkers(map, lng = NULL, lat = NULL, layerId = NULL,
  group = NULL, icon = NULL, popup = NULL, popupOptions = NULL,
  label = NULL, labelOptions = NULL, options = leaflet::markerOptions(),
  clusterOptions = NULL, clusterId = NULL,
  data = leaflet::getMapData(map))

Arguments

...

icons created from makePulseIcon()

x

icons

i

offset

color

Color of the icon

iconSize

Size of Icon in Pixels.

animate

To animate the icon or not, defaults to TRUE.

heartbeat

Interval between each pulse in seconds.

map

the map to add pulse Markers to.

lng

a numeric vector of longitudes, or a one-sided formula of the form ~x where x is a variable in data; by default (if not explicitly provided), it will be automatically inferred from data by looking for a column named lng, long, or longitude (case-insensitively)

lat

a vector of latitudes or a formula (similar to the lng argument; the names lat and latitude are used when guessing the latitude column from data)

layerId

the layer id

group

the name of the group the newly created layers should belong to (for clearGroup and addLayersControl purposes). Human-friendly group names are permitted--they need not be short, identifier-style names. Any number of layers and even different types of layers (e.g. markers and polygons) can share the same group name.

icon

the icon(s) for markers;

popup

a character vector of the HTML content for the popups (you are recommended to escape the text using htmlEscape() for security reasons)

popupOptions

options for popup

label

a character vector of the HTML content for the labels

labelOptions

A Vector of labelOptions to provide label options for each label. Default NULL

options

a list of extra options for tile layers, popups, paths (circles, rectangles, polygons, ...), or other map elements

clusterOptions

if not NULL, markers will be clustered using https://github.com/Leaflet/Leaflet.markercluster; you can use markerClusterOptions() to specify marker cluster options

clusterId

the id for the marker cluster layer

data

the data object from which the argument values are derived; by default, it is the data object provided to leaflet() initially, but can be overridden

Examples

iconSet = pulseIconList(
  red = makePulseIcon(color='#ff0000'),
  blue = makePulseIcon(color='#0000ff')
)

iconSet[c('red', 'blue')]

#> $red
#> $color
#> [1] "#ff0000"
#> 
#> $iconSize
#> [1] 12
#> 
#> $animate
#> [1] TRUE
#> 
#> $heartbeat
#> [1] 1
#> 
#> attr(,"class")
#> [1] "leaflet_pulse_icon"
#> 
#> $blue
#> $color
#> [1] "#0000ff"
#> 
#> $iconSize
#> [1] 12
#> 
#> $animate
#> [1] TRUE
#> 
#> $heartbeat
#> [1] 1
#> 
#> attr(,"class")
#> [1] "leaflet_pulse_icon"
#> 
#> attr(,"class")
#> [1] "leaflet_pulse_icon_set"
#>