Options for TiledMapLayer.

Options for TiledMapLayer.

tiledMapLayerOptions(correctZoomLevels = TRUE, zoomOffsetAllowance = 0.1,
  proxy = NULL, useCors = TRUE, token = NULL, tileOptions = NULL)

Arguments

correctZoomLevels
Correct Zoom levels.

zoomOffsetAllowance
If correctZoomLevels is enabled this controls the amount of tolerance for the difference at each scale level for remapping tile levels.

proxy
URL of an ArcGIS API for JavaScript proxy or ArcGIS Resource Proxy to use for proxying requests.

useCors
Dictates if the service should use CORS when making GET requests.

token
Use this token to authenticate all calls to the service.

tileOptions
Other options for tile layer. You can pass tileOptions().