Adds a Toolbar to draw shapes/points on the map.

Adds a Toolbar to draw shapes/points on the map.

Removes the draw toolbar

addDrawToolbar(map, targetLayerId = NULL, targetGroup = NULL,
  position = c("topleft", "topright", "bottomleft", "bottomright"),
  polylineOptions = drawPolylineOptions(),
  polygonOptions = drawPolygonOptions(),
  circleOptions = drawCircleOptions(),
  rectangleOptions = drawRectangleOptions(),
  markerOptions = drawMarkerOptions(), editOptions = FALSE)

removeDrawToolbar(map, clearFeatures = FALSE)

Arguments

map

The map widget.

targetLayerId

An optional layerId of a GeoJSON/TopoJSON layer whose features need to be editable. Used for adding a GeoJSON/TopoJSON layer and then editing the features using the draw plugin.

targetGroup

An optional group name of a Feature Group whose features need to be editable. Used for adding shapes(markers,lines,polygons) and then editing them using the draw plugin. You can either set layerId or group or none but not both.

position

The position where the toolbar should appear.

polylineOptions

See drawPolylineOptions(). Set to FALSE to disable polyline drawing.

polygonOptions

See drawPolygonOptions(). Set to FALSE to disable polygon drawing.

circleOptions

See drawCircleOptions(). Set to FALSE to disable circle drawing.

rectangleOptions

See drawRectangleOptions(). Set to FALSE to disable rectangle drawing.

markerOptions

See drawMarkerOptions(). Set to FALSE to disable marker drawing.

editOptions

By default editing is disable. To enable editing pass editToolbarOptions().

clearFeatures

whether to clear the map of drawn features.