Add a webgl heatmap

Add a webgl heatmap

removes the webgl heatmap

clears the webgl heatmap

addWebGLHeatmap(map, lng = NULL, lat = NULL, intensity = NULL,
  layerId = NULL, group = NULL, size = "30000", units = "m",
  opacity = 1, gradientTexture = NULL, alphaRange = 1,
  data = leaflet::getMapData(map))

removeWebGLHeatmap(map, layerId)

clearWebGLHeatmap(map)

Arguments

map

the map to add pulse Markers to.

lng

a numeric vector of longitudes, or a one-sided formula of the form ~x where x is a variable in data; by default (if not explicitly provided), it will be automatically inferred from data by looking for a column named lng, long, or longitude (case-insensitively)

lat

a vector of latitudes or a formula (similar to the lng argument; the names lat and latitude are used when guessing the latitude column from data)

intensity

intensity of the heat. A vector of numeric values or a formula.

layerId

the layer id

group

the name of the group the newly created layers should belong to (for clearGroup and addLayersControl purposes). Human-friendly group names are permitted--they need not be short, identifier-style names. Any number of layers and even different types of layers (e.g. markers and polygons) can share the same group name.

size

in meters or pixels

units

either 'm' or 'px'

opacity

for the canvas element

gradientTexture

image url or image

alphaRange

adjust transparency by changing to value between 0 and 1

data

the data object from which the argument values are derived; by default, it is the data object provided to leaflet() initially, but can be overridden